weapons breaking blocks fires an event to ensure they have permission to do so

+++ src/main/java/com/onewhohears/dscombat/entity/weapon/EntityWeapon.java
@@ -14,6 +14,7 @@ import com.onewhohears.onewholibs.util.UtilEntity;
 
 import com.onewhohears.onewholibs.util.UtilParse;
 import net.minecraft.client.Minecraft;
+import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.protocol.Packet;
@@ -25,9 +26,11 @@ import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.projectile.Projectile;
 import net.minecraft.world.level.ClipContext;
 import net.minecraft.world.level.ClipContext.Fluid;
+import net.minecraft.world.level.GameRules;
 import net.minecraft.world.level.GameType;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
@@ -37,8 +40,10 @@ import net.minecraft.world.phys.HitResult;
 import net.minecraft.world.phys.Vec3;
 import net.minecraft.world.scores.Team;
 import net.minecraftforge.common.ForgeHooks;
+import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.entity.IEntityAdditionalSpawnData;
 import net.minecraftforge.event.ForgeEventFactory;
+import net.minecraftforge.event.level.BlockEvent;
 import net.minecraftforge.network.NetworkHooks;
 import net.minecraftforge.network.PacketDistributor;
 
@@ -223,13 +228,21 @@ public abstract class EntityWeapon<T extends WeaponStats> extends Projectile imp
 		//System.out.println("BULLET HIT "+result.getBlockPos());
 		if (canBreakFragileBlocks()) {
 			BlockState state = getLevel().getBlockState(result.getBlockPos());
-			if (state.is(ModTags.Blocks.FRAGILE)) {
-				//ForgeHooks.onBlockBreakEvent(getLevel(), )
+			if (state.is(ModTags.Blocks.FRAGILE) && hasPermissionToBreakBlock(result.getBlockPos(), state)) {
 				getLevel().destroyBlock(result.getBlockPos(), true, this);
 				return;
 			}
         }
 		kill();
+	}
+
+	protected boolean hasPermissionToBreakBlock(BlockPos pos, BlockState state) {
+		if (getOwner() instanceof Player player) {
+			BlockEvent.BreakEvent event = new BlockEvent.BreakEvent(getLevel(), pos, state, player);
+			MinecraftForge.EVENT_BUS.post(event);
+			return !event.isCanceled();
+		}
+		return getLevel().getGameRules().getBoolean(GameRules.RULE_MOBGRIEFING);
 	}
 	
 	@Override
